ovl: fail on invalid uid/gid mapping at copy up
If st_uid/st_gid doesn't have a mapping in the mounter's user_ns, then
copy-up should fail, just like it would fail if the mounter task was doing
the copy using "cp -a".

There's a corner case where the "cp -a" would succeed but copy up fail: if
there's a mapping of the invalid uid/gid (65534 by default) in the user
namespace.  This is because stat(2) will return this value if the mapping
doesn't exist in the current user_ns and "cp -a" will in turn be able to
create a file with this uid/gid.

This behavior would be inconsistent with POSIX ACL's, which return -1 for
invalid uid/gid which result in a failed copy.

For consistency and simplicity fail the copy of the st_uid/st_gid are
invalid.

Bug 4108481
Bug 4125613
Bug 4139235

Change-Id: I9834808d7496694f3d0a2aa587c6122d34ae3db4
Fixes: 459c7c565ac3 ("ovl: unprivieged mounts")
Signed-off-by: Miklos Szeredi <mszeredi@redhat.com>
Reviewed-by: Christian Brauner <brauner@kernel.org>
Reviewed-by: Seth Forshee <sforshee@kernel.org>
Reviewed-on: https://git-master.nvidia.com/r/c/linux-4.9/+/2901445
Tested-by: Dmitry Pervushin <dpervushin@nvidia.com>
Reviewed-by: Dmitry Pervushin <dpervushin@nvidia.com>
(cherry picked from commit 813f5f7f7982ddc840b10bec9150611f70cf6ba7)
Reviewed-on: https://git-master.nvidia.com/r/c/linux-4.9/+/2913010
GVS: Gerrit_Virtual_Submit <buildbot_gerritrpt@nvidia.com>