bluetooth: Add cmd validity checks at the start of hci_sock_ioctl()
Previously, channel open messages were always sent to monitors on the first
ioctl() call for unbound HCI sockets, even if the command and arguments
were completely invalid. This can leave an exploitable hole with the abuse
of invalid ioctl calls.

This commit hardens the ioctl processing logic by first checking if the
command is valid, and immediately returning with an ENOIOCTLCMD error code
if it is not. This ensures that ioctl calls with invalid commands are free
of side effects, and increases the difficulty of further exploitation by
forcing exploitation to find a way to pass a valid command first.

Bug 4097937
Bug 4113823
Bug 4139239

Change-Id: Ibe1980ba39c65ca4639d26886bee19152b4d5c65
Signed-off-by: Ruihan Li <lrh2000@pku.edu.cn>
Co-developed-by: Marcel Holtmann <marcel@holtmann.org>
Signed-off-by: Marcel Holtmann <marcel@holtmann.org>
Reviewed-on: https://git-master.nvidia.com/r/c/linux-4.9/+/2902024
Tested-by: Dmitry Pervushin <dpervushin@nvidia.com>
Reviewed-by: Dmitry Pervushin <dpervushin@nvidia.com>
(cherry picked from commit b21ca1dd0a06cbcdccd4b12b908383b315bcca1b)
Reviewed-on: https://git-master.nvidia.com/r/c/linux-4.9/+/2913012
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>